Rename TransactionPool to TxPool
